refactor(canvas): split the grid surface out of GridCanvasView
Move the drawable surface — the lattice, the cards, and the pointer state
machine behind draw, move and resize — into GridSurface, with the empty-grid
hint as its own component. GridCanvasView keeps the canvas's data, toolbar
and chat panel, and drops from ~390 to ~200 lines, under react-doctor's
no-giant-component threshold.

The surface owns its ref, measurement and drag hook, so the grid definition
reaches useGridDrag after the loading guard and the DEFAULT_GRID fallback
goes with it.
